feat: main page slider
