+++ src/app/components/filters.ts
@@ -1,7 +1,7 @@
 import getCategories from '../api/category/getCategories';
 import getProductsTypes from '../api/products-types/getProductsTypes';
 import { catalogQueryParams, productLimit } from '../state/state';
-import { Category, ProductType, QueryParam } from '../types/types';
+import { Category, ProductType } from '../types/types';
 import { sorterParametrs } from './constants';
 import { createCheckBoxElement, createElement, createInputElement } from './utils';
 
@@ -170,10 +170,10 @@ class Filters {
     search.value = '';
 
     localStorage.removeItem('search_products');
-    catalogQueryParams.clear();
-    catalogQueryParams.forEach((param: QueryParam) => {
-      if (param.key !== 'sidebar') catalogQueryParams.delete(param.key);
-    });
+
+    for (const param of catalogQueryParams.keys()) {
+      if (param !== 'sidebar') catalogQueryParams.delete(param);
+    }
 
     productLimit.limit = 12;
   }
